requirements/test.txt
factory_boy == 2.9.2
flake8 == 3.4.1
mock == 2.0.0 ; python_version < '3.3'
pytest == 3.2.1
pytest-django == 3.1.2
